pushed feedback changes
