+++ src/components/modals/termsAndConditions.tsx
@@ -4,8 +4,8 @@ import SignaturePad from "react-signature-canvas";
 import React, { useRef, useState, useEffect } from "react";
 import { Boat, Booking, BookingFormData } from "@/models/models";
 import moment from "moment";
-import { toast } from "react-toastify";
 import { uploadSignatureImage } from "@/services/googleDrive.service";
+import { convertCanvasToImage } from "@/services/utils";
 
 export default function TermsAndConditions({
   isOpen,
@@ -72,43 +72,11 @@ export default function TermsAndConditions({
     maximumDepartureTime()
   }, [bookingInfo]);
 
-  const getImage = async (
-    {
-      canvas,
-      mime,
-    }: {
-      canvas: HTMLCanvasElement,
-      mime: string,
-    }
-  ) => {
-    return new Promise<Blob | null>(resolve => {
-      const tmpCanvas = document.createElement("canvas");
-      // Use the original canvas's dimensions
-      tmpCanvas.width = canvas.width;
-      tmpCanvas.height = canvas.height;
-
-      const ctx = tmpCanvas.getContext("2d");
-      if (!ctx) {
-        return resolve(null); // Safely handle the case where context could not be obtained
-      }
-
-      // Set the background to white
-      ctx.fillStyle = "#FFFFFF"; // Define white color
-      ctx.fillRect(0, 0, tmpCanvas.width, tmpCanvas.height); // Fill the canvas with white
-
-      // Draw the original canvas onto the temporary canvas without resizing
-      ctx.drawImage(canvas, 0, 0);
-
-      // Convert the temporary canvas to a Blob
-      tmpCanvas.toBlob(blob => resolve(blob), mime);
-    });
-  };
-
   const getSignatureImage = async () => {
     if (sigPad.current) {
       const mime = "image/jpeg"
       const canvas = sigPad.current.getTrimmedCanvas();
-      const image = await getImage({ canvas, mime })
+      const image = await convertCanvasToImage({ canvas, mime })
 
       const response = await uploadSignatureImage(image as File);
       if (!response.id) {
