+++ src/i18n/en.json
@@ -109,7 +109,10 @@
     "ASB-409": "This booking is already submitted.",
     "CSC-503": "Contract already signed.",
     "maximum_number_250": "The maximum limit for sending messages at once is 250 customers.",
-    "failed_to_send_message_to": "Failed to send message to this number"
+    "failed_to_send_message_to": "Failed to send message to this number",
+    "error_booking_details": "Booking details not found",
+    "error_boat_details": "Boat details not found",
+    "error_captain_details": "Captain details not found"
   },
   "title": {
     "feed_back_form": "Treatment Received Feedback Form",
